requirements.txt
lancedb
pyarrow
numpy
google-generativeai
pandas
